further modification to merchant pages
